refactor: card primitive
