+++ package.json
@@ -6,6 +6,7 @@
     "@hashicorp/mktg-global-styles": "^4.3.1",
     "@hashicorp/mktg-logos": "^1.3.2",
     "@hashicorp/sentinel-embedded": "^0.0.14",
+    "@swc-node/jest": "^1.5.5",
     "classnames": "^2.3.1",
     "formik": "^2.2.9",
     "framer-motion": "^6.3.11",
@@ -31,7 +32,6 @@
     "@hashicorp/platform-types": "^0.3.0",
     "@hashicorp/platform-util": "0.1.0",
     "@octokit/core": "^3.5.1",
-    "@swc-node/jest": "^1.5.2",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^8.0.0",
