+++ packages/card/docs.mdx
@@ -108,99 +108,6 @@ A component used to promote and link to marketing pages, always using a descript
 </div>
 
 <h3 className="g-type-label" style={{ color: 'var(--gray-3)' }}>
-  Using composable children
-</h3>
-
-<div
-  style={{
-    display: 'grid',
-    gridTemplateColumns: '1fr 1fr 1fr',
-    gap: '32px',
-    marginBottom: '32px',
-  }}
->
-  <Card link="https://hashicorp.com">
-    <Card.Thumbnail
-      src="https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png"
-      alt="HashiConf Europe 2022 Recap"
-    />
-    <Card.Content>
-      <Card.Heading>
-        {
-          'Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…'
-        }
-      </Card.Heading>
-    </Card.Content>
-  </Card>
-  <Card link="https://hashicorp.com">
-    <Card.Content>
-      <Card.Meta items={['August 15, 2022', 'Category']} />
-      <Card.Heading>
-        {
-          'Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…'
-        }
-      </Card.Heading>
-      <Card.Description>
-        {
-          'Body small - 4 lines with character limit of 170. Leo mauris fermentum pharetra blandit tellus euismod pharetra. Duis egestas volutpat dolor…'
-        }
-      </Card.Description>
-    </Card.Content>
-  </Card>
-  <Card link="https://hashicorp.com">
-    <Card.Content>
-      <Card.Heading>
-        {
-          'Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…'
-        }
-      </Card.Heading>
-      <p style={{ fontSize: '16px' }}>{'Example arbitrary content'}</p>
-    </Card.Content>
-  </Card>
-</div>
-
-<h3 className="g-type-label" style={{ color: 'var(--gray-3)' }}>
-  Using different heading elements
-</h3>
-
-<div
-  style={{
-    display: 'grid',
-    gridTemplateColumns: '1fr 1fr 1fr',
-    gap: '32px',
-    marginBottom: '32px',
-  }}
->
-  <Card link="https://hashicorp.com">
-    <Card.Content>
-      <Card.Heading>
-        {
-          'Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…'
-        }
-      </Card.Heading>
-    </Card.Content>
-  </Card>
-  <Card link="https://hashicorp.com">
-    <Card.Content>
-      <Card.Heading as="h3">
-        {
-          'Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…'
-        }
-      </Card.Heading>
-    </Card.Content>
-  </Card>
-  <Card link="https://hashicorp.com">
-    <Card.Content>
-      <Card.Heading as="h4">
-        {
-          'Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…'
-        }
-      </Card.Heading>
-    </Card.Content>
-  </Card>
-</div>
-
-<h3 className="g-type-label" style={{ color: 'var(--gray-3)' }}>
   Use Case Cards with Product Badges
 </h3>
 
@@ -215,72 +122,34 @@ A component used to promote and link to marketing pages, always using a descript
   <Card
     link="https://hashicorp.com"
     thumbnail={{
-      src: "https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png",
-      alt: "HashiConf Europe 2022 Recap"
+      src: 'https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png',
+      alt: 'HashiConf Europe 2022 Recap',
     }}
-    meta={[
-      'USE CASE'
-    ]}
+    meta={['USE CASE']}
     heading="Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…"
-    productBadges={[
-        'waypoint',
-        'consul',
-      ]}
+    productBadges={['waypoint', 'consul']}
     description="Body small - 4 lines with character limit of 170. Leo mauris fermentum pharetra blandit tellus euismod pharetra. Duis egestas volutpat dolor…"
   />
   <Card
     link="https://hashicorp.com"
     thumbnail={{
-      src: "https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png",
-      alt: "HashiConf Europe 2022 Recap"
+      src: 'https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png',
+      alt: 'HashiConf Europe 2022 Recap',
     }}
-    meta={[
-      'USE CASE'
-    ]}
+    meta={['USE CASE']}
     heading="Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…"
-    productBadges={[
-        'terraform',
-        'consul',
-      ]}
+    productBadges={['terraform', 'consul']}
     description="Body small - 4 lines with character limit of 170. Leo mauris fermentum pharetra blandit tellus euismod pharetra. Duis egestas volutpat dolor…"
   />
   <Card
     link="https://hashicorp.com"
     thumbnail={{
-      src: "https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png",
-      alt: "HashiConf Europe 2022 Recap"
+      src: 'https://www.datocms-assets.com/19447/1648680074-screenshot-2022-03-31-at-00-40-47.png',
+      alt: 'HashiConf Europe 2022 Recap',
     }}
-    meta={[
-      'USE CASE'
-    ]}
+    meta={['USE CASE']}
     heading="Display 5 - 4 lines with character limit of 130. Leo mauris fermentum pharetra blandit tellus euismod…"
-    productBadges={[
-      'nomad',
-      'vault',
-      'terraform'
-    ]}
+    productBadges={['nomad', 'vault', 'terraform']}
     description="Body small - 4 lines with character limit of 170. Leo mauris fermentum pharetra blandit tellus euismod pharetra. Duis egestas volutpat dolor…"
   />
 </div>
-
-## Base Props
-
-<PropsTable props={componentProps.baseProps} />
-
-## Sub-component-specific Props
-
-### `<Thumbnail />`
-
-<PropsTable props={componentProps.thumbnailProps} />
-
-### `<Meta />`
-
-<PropsTable props={componentProps.metaProps} />
-
-### `<Heading />`
-
-<PropsTable props={componentProps.headingProps} />
-
-### `<Description />`
-
-<PropsTable props={componentProps.descriptionProps} />
