+++ packages/card/index.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import Card from '.'
+import Card from './'
 
 const defaultProps = {
   link: 'https://hashicorp.com',
@@ -60,96 +60,6 @@ describe('<Card />', () => {
     const { getByAltText } = render(<Card {...defaultProps} />)
 
     expect(getByAltText(defaultProps.thumbnail.alt).tagName).toBe('IMG')
-  })
-
-  // Children render tests
-
-  it('should render provided Thumbnail sub-component children', () => {
-    const { getByAltText } = render(
-      <Card link={defaultProps.link}>
-        <Card.Thumbnail {...defaultProps.thumbnail} />
-      </Card>
-    )
-
-    expect(getByAltText(defaultProps.thumbnail.alt).tagName).toBe('IMG')
-  })
-
-  it('should render provided Meta sub-component children', () => {
-    render(
-      <Card link={defaultProps.link}>
-        <Card.Meta items={defaultProps.meta} />
-      </Card>
-    )
-
-    const metaElement = screen.getByTestId('wpl-card-meta')
-
-    defaultProps.meta.forEach((item) => {
-      expect(metaElement).toContainElement(screen.getByText(item))
-    })
-  })
-
-  it('should render provided Heading sub-component children', () => {
-    render(
-      <Card link={defaultProps.link}>
-        <Card.Heading>{defaultProps.heading}</Card.Heading>
-      </Card>
-    )
-
-    screen.getByText(defaultProps.heading)
-  })
-
-  it('should render provided Description sub-component children', () => {
-    render(
-      <Card link={defaultProps.link}>
-        <Card.Description>{defaultProps.description}</Card.Description>
-      </Card>
-    )
-
-    screen.getByText(defaultProps.description)
-  })
-
-  it('should render arbitrary JSX children', () => {
-    const arbitraryText = 'Arbitrary paragraph text'
-    const arbitraryParagraph = <p>{arbitraryText}</p>
-
-    render(<Card link={defaultProps.link}>{arbitraryParagraph}</Card>)
-
-    screen.getByText(arbitraryText)
-  })
-
-  it('should render arbitrary text content', () => {
-    const arbitraryText = 'Arbitrary text'
-
-    render(<Card link={defaultProps.link}>{arbitraryText}</Card>)
-
-    screen.getByText(arbitraryText)
-  })
-
-  it('should not render props-based layout if children are provided', () => {
-    const arbitraryText = 'Arbitrary text'
-
-    render(
-      <Card link={defaultProps.link} heading={defaultProps.heading}>
-        {arbitraryText}
-      </Card>
-    )
-
-    screen.getByText(arbitraryText)
-    expect(screen.queryByText(defaultProps.heading)).not.toBeInTheDocument()
-  })
-
-  it('should render correct levels for the Heading sub-component via the as prop', () => {
-    render(
-      <Card link={defaultProps.link}>
-        <Card.Heading>As h2</Card.Heading>
-        <Card.Heading as="h3">As h3</Card.Heading>
-        <Card.Heading as="h4">As h4</Card.Heading>
-      </Card>
-    )
-
-    expect(screen.getByText('As h2').tagName).toBe('H2')
-    expect(screen.getByText('As h3').tagName).toBe('H3')
-    expect(screen.getByText('As h4').tagName).toBe('H4')
   })
 
   // Appearance tests
